Add Bucket Meters:
* `BucketTimer`: A timer that manages a set of timers that get updated
based on the bucket function supplied.

* `BucketCounter`: A distribution summary that manages a set a counters
that will be updated based on the bucket function supplied.

* `BucketDistributionSummary`: A distribution summary that manages a set
of distribution summaries that will be updated based on the bucket
function supplied.

A set of `BucketFunctions` is supplied:

* `age`: a function that maps ages to a set of buckets creating more
buckets towards the fresher end of the range.

* `ageBiasOld`: a function that maps ages to a set of buckets creating more
buckets towards the older end of the range.

* `latency`: a function that maps latencies to a set of buckets
creating more buckets towards the faster end of the range.

* `latencyBiasSlow`: a function that maps latencies to a set of buckets
creating more buckets towards the slower end of the range.

* `bytes`: a function that maps size values to a set of buckets using
powers of 2.

* `decimal`: a function that maps size values to a set of buckets using
powers of 10.