feat: add TypeScript declarations (#44)
* feat: add TypeScript declarations

This is broken up into separate files because some users point directly to a file.

Add tests to main Registry class.

* refactor: move declaration files before publish

Instead of pointing to a types dir copy files over to src so editors can pick up the correct location. I added the clean up and copy as separate steps so it can be run locally.

Also added tsd to test command.