remove references to unix domain sockets (#96)
The `unix_gram` module was removed from the standard library in 2011, for
Windows portability reasons, so this library will not support Unix Domain
Sockets.

https://github.com/nodejs/node/issues/29339

There is a third-party library `unix-dgram`, which offers this functionality,
but it introduces some complications, and we want this library to be as low-
friction as possible. Thus, we will not adopt this module.

https://github.com/bnoordhuis/node-unix-dgram

This state of affairs is similar to the Java thin-client, where the JVM does
not support Unix Domain Sockets until 16+:

https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/net/UnixDomainSocketAddress.html).